pyproject.toml
[project]
name = "tg-listener"
version = "0.1.0"
description = "Automated Telegram message listener with regex filtering and macOS Shortcuts integration"
readme = "README.md"

requires-python = ">=3.13"
dependencies = [
    "aiochannel>=1.3.0",
    "multipledispatch>=1.0.0",
    "python-dotenv>=1.1.1",
    "telethon>=1.41.2",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "mypy>=1.18.2",
    "ruff>=0.9.0",
    "wemake-python-styleguide>=1.4.0",
    "bandit==1.7.10",
    "semgrep==1.92.0",
    "codespell==2.3.0",
    "dotenv-linter>=0.7.0",
    "flake8>=7.3.0",
    "flake8-eradicate>=1.5.0",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]

[tool.mypy]
strict = true
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.ruff]
exclude = [
    ".junie",
    ".venv",
    "run",
]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "D203", # one-blank-line-before-class (conflicts with D211)
    "D213", # multi-line-summary-second-line (conflicts with D212)
    "COM812", # trailing-comma-missing (conflicts with formatter)
    "ISC001", # single-line-implicit-string-concatenation (conflicts with formatter)
    "CPY001", # missing-copyright-notice
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "PLR2004", "SLF001", "D100", "D101", "D102", "ANN201", "ANN001", "ARG002"]
"app/*" = ["I001", "F811"]

[tool.ruff.lint.pydocstyle]
convention = "google"
